+++ kafka-streams/kafka-streams-queryable-thrift/src/main/scala/com/twitter/finatra/streams/queryable/thrift/QueryableState.scala
@@ -7,9 +7,8 @@ import com.twitter.finatra.streams.query.{
   QueryableFinatraKeyValueStore,
   QueryableFinatraWindowStore
 }
-import com.twitter.finatra.streams.transformer.domain.TimeWindowed
+import com.twitter.util.Duration
 import org.apache.kafka.common.serialization.Serde
-import org.apache.kafka.streams.state.QueryableStoreTypes
 
 trait QueryableState extends StaticPartitioning {
   protected val currentShard: Flag[Int] = flag[Int]("kafka.current.shard", "")
@@ -20,38 +19,46 @@ trait QueryableState extends StaticPartitioning {
     primaryKeySerde: Serde[PK]
   ): QueryableFinatraKeyValueStore[PK, K, V] = {
     new QueryableFinatraKeyValueStore[PK, K, V](
-      kafkaStreams.store(storeName, QueryableStoreTypes.keyValueStore[K, V]),
-      primaryKeySerde,
-      numApplicationInstances(),
-      numQueryablePartitions(),
-      currentShard()
-    )
-  }
-
-  protected def queryableFinatraWindowStore[K, V](
-    storeName: String,
-    primaryKeySerde: Serde[K]
-  ): QueryableFinatraWindowStore[K, V] = {
-    new QueryableFinatraWindowStore[K, V](
-      kafkaStreams.store(storeName, QueryableStoreTypes.keyValueStore[TimeWindowed[K], V]),
-      primaryKeySerde,
-      numApplicationInstances(),
-      numQueryablePartitions(),
-      currentShard()
-    )
-  }
-
-  protected def queryableFinatraCompositeWindowStore[K, PK, SK, V](
-    storeName: String,
-    primaryKeySerde: Serde[PK]
-  ): QueryableFinatraCompositeWindowStore[K, PK, SK, V] = {
-    new QueryableFinatraCompositeWindowStore[K, PK, SK, V](
-      kafkaStreams,
       storeName,
       primaryKeySerde,
       numApplicationInstances(),
       numQueryablePartitions(),
-      currentShard()
-    )
+      currentShard())
+  }
+
+  protected def queryableFinatraWindowStore[K, V](
+    storeName: String,
+    windowSize: Duration,
+    primaryKeySerde: Serde[K]
+  ): QueryableFinatraWindowStore[K, V] = {
+    new QueryableFinatraWindowStore[K, V](
+      storeName,
+      windowSize = windowSize,
+      keySerde = primaryKeySerde,
+      numShards = numApplicationInstances(),
+      numQueryablePartitions = numQueryablePartitions(),
+      currentShardId = currentShard())
+  }
+
+  @deprecated("Use queryableFinatraWindowStore without a windowSize", "1/7/2019")
+  protected def queryableFinatraWindowStore[K, V](
+    storeName: String,
+    primaryKeySerde: Serde[K]
+  ): QueryableFinatraWindowStore[K, V] = {
+    queryableFinatraWindowStore(storeName, null, primaryKeySerde)
+  }
+
+  protected def queryableFinatraCompositeWindowStore[PK, SK, V](
+    storeName: String,
+    windowSize: Duration,
+    primaryKeySerde: Serde[PK]
+  ): QueryableFinatraCompositeWindowStore[PK, SK, V] = {
+    new QueryableFinatraCompositeWindowStore[PK, SK, V](
+      storeName,
+      windowSize = windowSize,
+      primaryKeySerde = primaryKeySerde,
+      numShards = numApplicationInstances(),
+      numQueryablePartitions = numQueryablePartitions(),
+      currentShardId = currentShard())
   }
 }